+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## HEAD | 1.3.1 | 05.03.2018
+* cleanup gulpfile, add variables to watch-task
+
 ## 1.3.0 | 05.03.2018 | «Vendorizr»
 * add `gulp-vendors.json` for compiling/bundling of vendor-requirements
 * add configuration variables for assets and vendor paths